requirements.txt
discord.py==1.7.2
PyAutoGUI==0.9.52